display: carve encode/pool lease.rs (pure move)
